gynaecological_cancer/requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.3
matplotlib==3.9.0
pillow==10.2.0
tqdm==4.65.2
